Add plugin version to samples so that they can be executed
Users should be able to try out samples by checking out the source code contained in this repository. A version is required by the plugin DSL so that it can resolve the plugin from the Gradle plugin portal.

The documentation tests always use the injected plugin code that is currently under development by setting `withPluginClasspath()`. The user guide includes snippets from the samples. The task generating the HTML version of the user guide replace the plugin version with the next version to be released. This will ensure that the plugin version is correct after a new version has been released.

As part of the release process, the developer has to manually update the plugin version in the samples code.