Add Play 2.8.x compatibility
Play's 2.8 changelog does not mention much actually needing
to change from the build pipeline perspective, so an easy
string replacement for the integration test fixtures to verify
and test should be enough to verify.

Based on the previous PR from gradle/playframework#133, this
expands the changes to ensure the integration tests pass as
well.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>